requirements.txt

# Django
Django==3.1
django-environ==0.4.5
psycopg2-binary==2.8.6
gunicorn==20.1.0

# DRF
djangorestframework==3.11.0
django-cors-headers==3.5.0
drf-nested-routers==0.92.0
drf-extensions==0.7.0

# JWT
djangorestframework_simplejwt==4.6.0

# Custom Choices
django-model-utils==4.0.0

# Swagger
drf-yasg==1.17.1
flex==6.14.1
swagger-spec-validator==2.7.3

# Development
ipdb==0.13.5

# Filters
django-filter==2.4.0

# Tests
coverage==5.5

# PEP8
black==20.8b1

# Date
python-dateutil==2.8.1

# Middleware
django-xff==1.3.0
